+++ src/core_graphics/mod.rs
@@ -7,3 +7,7 @@
 
 #[link(name = "CoreGraphics", kind = "framework")]
 extern "C" {}
+
+mod geometry;
+
+pub use geometry::*;